changed to float

+++ c_src/calculate.c
@@ -2,19 +2,19 @@
 #include <math.h>
 #include "calculate.h"
 
-double add(double num1, double num2)
+float add(float num1, float num2)
 {
     return num1 + num2;
 }
-double subtract(double num1, double num2)
+float subtract(float num1, float num2)
 {
     return num1 - num2;
 }
-double multiply(double num1, double num2)
+float multiply(float num1, float num2)
 {
     return num1 * num2;
 }
-double divide(double num1, double num2)
+float divide(float num1, float num2)
 {
     if (num2 != 0.0)
     {
