+++ c_src/calculate.h
@@ -1,6 +1,6 @@
 #define CALCULATE_H
 
-double add(double num1, double num2);
-double subtract(double num1, double num2);
-double multiply(double num1, double num2);
-double divide(double num1, double num2);
+float add(float num1, float num2);
+float subtract(float num1, float num2);
+float multiply(float num1, float num2);
+float divide(float num1, float num2);
